txml: BaseXMLHandler: allow users to set callbacks for specific attributes
Instead of making the Parser class call the item callbacks,
BaseXMLHandler will do that as requested. This will allow use to be
notified of different items using the same Parser object (e.g. on the
paging mode, where we want to be notified about page data, and also
about items inside pages).

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>