Simplify TwitterMonitor's API.
This drops the `refreshFilters` method and makes `__init__` take
parameters similar to `TwitterFeed`. Instead of an object with a
`onEntry` method, a simple callable is now taken as the consumer
of incoming tweets (now called delegate). The errors raised by `connect`
are also split over multiple exception classes.

--HG--
branch : streaming-monitor